Implement basic 8points based initialisation
For the moment only for two cameras, PnP should still be added for the
other cameras and Triangulation should be used for the missing points.

Note: The model could also be extended to estimate the depth of the
initial points in the initial camera frame using more than one
additional camera.

+++ gui/sparsealignementviewer.cpp
@@ -305,7 +305,7 @@ void SparseAlignementViewer::paintGL() {
 
 					if (cam != nullptr) {
 						QSize s = cam->imSize();
-						qreal aspect_ratio = s.width()/s.height();
+						qreal aspect_ratio = static_cast<qreal>(s.width())/static_cast<qreal>(s.height());
 						if (aspect_ratio > 1) {
 							camScale.scale(1., 1./aspect_ratio);
 						} else {
